Improved exception handling
The application will now stop running after the user accepts the error message notifying them that the tp_mod.ini was not found.

This prevents additional unhandled exception messages from appearing and removes the need for multiple functions to have exception handling. Instead, a new function validates the presence of necessary files before the rest of the code runs.