Added persistent user settings
TMM now generates tmm_config.ini, which stores the path for the Tres CE exe set by the user.

The user is now prompted to set the Tres CE exe path if they press the Launch game button and a path is not already set or the path is no longer valid.

Updated version number.

Removed obsolete code.

Changed "check_installed_mods()" to "get.installed.mods()" for clarity and to match coding standards.